Move around generation-related bits
