+++ MetaMark/src/typescript_ast/TypeExpressionTerm.ts
@@ -2,4 +2,7 @@ import { TypeReferenceTerm } from "./TypeReferenceTerm";
 import { UnionTypeTerm } from "./UnionTypeTerm";
 import { StringLiteralTerm } from "./StringLiteralTerm";
 
-export type TypeExpressionTerm = TypeReferenceTerm | UnionTypeTerm | StringLiteralTerm;
+export type TypeExpressionTerm =
+    | TypeReferenceTerm
+    | UnionTypeTerm
+    | StringLiteralTerm;
